Post implementation

+++ backend/src/main/java/com/example/postgres/repository/ChannelRepository.java
@@ -3,5 +3,5 @@ package com.example.postgres.repository;
 import com.example.postgres.classes.Channel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChannelRepository extends JpaRepository<Channel, Integer> {
+public interface ChannelRepository extends JpaRepository<Channel, Long> {
 }
